src/requirements.txt
aiohttp>=3.9.0
beautifulsoup4==4.12.2
Jinja2==3.1.4
openai==1.6.1
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.32.2
sanic==23.12.0
sanic-ext==23.12.0
sanic_session==0.8.0
sanic-routing==23.12.0
htmlmin==0.1.12
tldextract==5.1.1
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic>=2.6.0
redis>=5.0.1
python-Levenshtein==0.25.0
aiograpi==0.0.3
Pillow>=10.4.0
thefuzz==0.22.1
msgspec==0.18.6
miniopy-async==1.20.1